Visualization plot for parameter optimzation
This function takes two inputs: param_values and performance_metrics. param_values is a list of dictionaries, where each dictionary contains the parameter values for one iteration of the optimization process. performance_metrics is a list of performance metrics, where each value corresponds to the performance of the model for one iteration of the optimization process.

The function then creates a sub-plot for each parameter in param_values, and adds the parameter values to each sub-plot. It also creates a separate plot for the performance metric and adds the performance metrics to that plot.